Move all test constants to seperate file.
