+++ ion/services/sa/instrument/test/test_agent_launch_ops.py
@@ -2,7 +2,7 @@ from interface.services.cei.iprocess_dispatcher_service import ProcessDispatcher
 from interface.services.dm.idataset_management_service import DatasetManagementServiceClient
 
 #from pyon.ion.endpoint import ProcessRPCClient
-from ion.agents.instrument.test.test_instrument_agent import DRV_URI_GOOD
+from ion.agents.instrument.test.agent_test_constants import DRV_URI_GOOD
 from ion.agents.port.port_agent_process import PortAgentProcessType, PortAgentType
 from ion.services.sa.instrument.agent_configuration_builder import PlatformAgentConfigurationBuilder, InstrumentAgentConfigurationBuilder
 from ion.util.enhanced_resource_registry_client import EnhancedResourceRegistryClient
