+++ ion/services/sa/process/test/test_data_process_with_lookup_tables.py
@@ -23,7 +23,7 @@ from pyon.public import RT, PRED
 from pyon.util.context import LocalContextMixin
 from pyon.util.int_test import IonIntegrationTestCase
 from ion.services.dm.utility.granule_utils import time_series_domain
-from ion.agents.instrument.test.test_instrument_agent import DRV_URI_GOOD
+from ion.agents.instrument.test.agent_test_constants import DRV_URI_GOOD
 
 import base64
