+++ ion/services/sa/test/test_assembly.py
@@ -29,7 +29,7 @@ from ion.services.sa.instrument.flag import KeywordFlag
 from ion.services.dm.utility.granule_utils import time_series_domain
 
 from ion.agents.port.port_agent_process import PortAgentType
-from ion.agents.instrument.test.test_instrument_agent import DRV_URI_GOOD
+from ion.agents.instrument.test.agent_test_constants import DRV_URI_GOOD
 from interface.services.dm.idataset_management_service import DatasetManagementServiceClient
 
 import string
